weather_icons
The weather icons file is a set of vector icons that represent different weather conditions provided by the OpenWeatherMap API. Each icon is designed to display a specific weather state such as sunny, cloudy, rainy, snowy, etc. These icons can be used to visually represent the weather information retrieved from the API in an intuitive and user-friendly way.

The weather icons file should include a set of vector graphics files that are easily scalable and can be used in different sizes and resolutions. The file format should be compatible with different platforms and frameworks, such as Android, iOS, and web development.